// src/core/transaction/types/transaction.types.ts
export enum TransactionType {
  ORDER_PAYMENT = 'ORDER_PAYMENT',
  SETTLEMENT = 'SETTLEMENT',
  WITHDRAWAL = 'WITHDRAWAL',
  REFUND = 'REFUND',
  REFUND_REVERSAL = 'REFUND_REVERSAL',
  WALLET_FUNDING = 'WALLET_FUNDING',
  PLATFORM_FEE = 'PLATFORM_FEE',
  PAYMENT_GATEWAY_FEE = 'PAYMENT_GATEWAY_FEE',
  CHARGEBACK = 'CHARGEBACK',
  ADJUSTMENT = 'ADJUSTMENT',
}

export enum TransactionStatus {
  PENDING = 'PENDING',
  PROCESSING = 'PROCESSING',
  SUCCESS = 'SUCCESS',
  FAILED = 'FAILED',
  REVERSED = 'REVERSED',
  CANCELLED = 'CANCELLED',
}

export enum TransactionFlow {
  CREDIT = 'CREDIT',
  DEBIT = 'DEBIT',
}